requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
asgiref==3.8.1
    # via django
black==24.8.0
    # via -r requirements/dev.in
click==8.1.7
    # via
    #   black
    #   uvicorn
django==4.2.16
    # via
    #   -r requirements/base.in
    #   django-stubs
    #   django-stubs-ext
django-stubs==4.2.7
    # via -r requirements/base.in
django-stubs-ext==5.1.0
    # via django-stubs
h11==0.14.0
    # via uvicorn
isort==5.13.2
    # via -r requirements/dev.in
markdown==3.7
    # via -r requirements/base.in
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
python-dotenv==1.0.1
    # via -r requirements/base.in
python-frontmatter==1.1.0
    # via -r requirements/base.in
python-slugify==8.0.4
    # via -r requirements/base.in
pyyaml==6.0.2
    # via python-frontmatter
sqlparse==0.5.1
    # via django
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via -r requirements/base.in
types-pytz==2024.2.0.20241003
    # via django-stubs
types-pyyaml==6.0.12.20240917
    # via django-stubs
typing-extensions==4.12.2
    # via
    #   django-stubs
    #   django-stubs-ext
uvicorn==0.31.0
    # via -r requirements/base.in
